moves HoverflyRuleConfigurer to hoverfly package
